requirements.txt
Flask==2.0.1
Flask-Cors==5.0.0
Flask-Login==0.5.0
Flask-Migrate==4.1.0
Flask-SocketIO==5.1.1
python-socketio==5.8.0
python-engineio==4.8.0
simple-websocket==1.1.0
python-dotenv==0.19.0
gunicorn==21.2.0
requests==2.31.0
Werkzeug==2.0.1
SQLAlchemy==1.4.1
alembic==1.14.1
APScheduler==3.11.0
email_validator==2.2.0
geographiclib==2.0
geopy==2.4.0
pg8000==1.30.5